requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
typer>=0.12.0
rich>=13.0.0
